extreme: refactor code
